% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_filteration.R
\name{deseq_filteration}
\alias{deseq_filteration}
\title{deseq filtration uses common filtration methods applied to post-processed RNA-Seq data
The goal is make my life easier by streamlining this process
Added bones each step is written out to excel spread sheet so that you can see your changes happen and show others
deseq_filtration will return a data.fram of the results and the name of the file you've chosen}
\usage{
deseq_filteration(tibble, file_name, pval_cutoff, rev = FALSE)
}
\description{
deseq filtration uses common filtration methods applied to post-processed RNA-Seq data
The goal is make my life easier by streamlining this process
Added bones each step is written out to excel spread sheet so that you can see your changes happen and show others
deseq_filtration will return a data.fram of the results and the name of the file you've chosen
}
