% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_filteration.R
\name{df_to_txt}
\alias{df_to_txt}
\title{df_to_txt was my first function I made by myself it gets a special seat.}
\usage{
df_to_txt(tibble)
}
\description{
df_to_txt was my first function I made by myself it gets a special seat.
}
