% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichGO.R
\name{enrichGO}
\alias{enrichGO}
\title{GO Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enrichment GO
categories after FDR control.}
\usage{
enrichGO(
  gene,
  OrgDb,
  keyType = "ENTREZID",
  ont = "MF",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500,
  readable = FALSE,
  pool = FALSE
)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{OrgDb}{OrgDb}

\item{keyType}{keytype of input gene}

\item{ont}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three.}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{readable}{whether mapping gene ID to gene Name}

\item{pool}{If ont='ALL', whether pool 3 GO sub-ontologies}
}
\value{
An \code{enrichResult} instance.
}
\description{
GO Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enrichment GO
categories after FDR control.
}
\examples{
\dontrun{
  data(geneList, package = "DOSE")
	de <- names(geneList)[1:100]
	yy <- enrichGO(de, 'org.Hs.eg.db', ont="BP", pvalueCutoff=0.01)
	head(yy)
}
}
\seealso{
\code{\link{enrichResult-class}}, \code{\link{compareCluster}}
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
\keyword{manip}
