% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichgo_unpack.R
\name{enrichgo_unpack}
\alias{enrichgo_unpack}
\title{enrichgo_unpack description
unpacks the genes based on the specified Description term from the enrichGO results an return all the genes under the the term
along side their deseq2 results. This will prepare them for volcano plot visualization
arguments
deseq2_table: processed data with the following columns: entrezgene_id,log2FoldChange, pvalue, padjust (high reccommend these four columns be present)
enrichGO reslts from ClusterProfiler package, the the chosen description (chosen_desc), and the ensembl_dataset that will passed to the above fxn
This will prep your data for the Volcano plot visual function and you can export and report the genes
BIG SIDE NOTE: this will only pre your data for the volcano plot fxn and not thed dotplot, visual function
example inputs
enrichgo_result <- key@result
chosen_desc <- "stress response to metal ion"
ensemble_dataset <- "hsapiens_gene_ensembl"
deseq2_table <- object1
enrichgo_unpack(deseq2_table, enrichgo_result, chosen_desc, ensemble_dataset)}
\usage{
enrichgo_unpack(deseq2_table, enrichgo_result, chosen_desc, ensembl_dataset)
}
\description{
enrichgo_unpack description
unpacks the genes based on the specified Description term from the enrichGO results an return all the genes under the the term
along side their deseq2 results. This will prepare them for volcano plot visualization
arguments
deseq2_table: processed data with the following columns: entrezgene_id,log2FoldChange, pvalue, padjust (high reccommend these four columns be present)
enrichGO reslts from ClusterProfiler package, the the chosen description (chosen_desc), and the ensembl_dataset that will passed to the above fxn
This will prep your data for the Volcano plot visual function and you can export and report the genes
BIG SIDE NOTE: this will only pre your data for the volcano plot fxn and not thed dotplot, visual function
example inputs
enrichgo_result <- key@result
chosen_desc <- "stress response to metal ion"
ensemble_dataset <- "hsapiens_gene_ensembl"
deseq2_table <- object1
enrichgo_unpack(deseq2_table, enrichgo_result, chosen_desc, ensemble_dataset)
}
