% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_visuals.R
\name{gsea_dotplot}
\alias{gsea_dotplot}
\title{arguments
Required columns: GSEA pvalue, GSEA Term, GSEA Count column
organize by Count or pvalue GSEA results (org = 'Count', org = 'pvalue')
num_disp
GSEA pvalue filter
graph title (title)
file_name write out
output visual dotplot of GSEA enrichment
input post GSEA}
\usage{
gsea_dotplot(
  gsea_results,
  gspval_cutoff,
  org_by,
  num_disp,
  graph_title,
  file_name
)
}
\description{
arguments
Required columns: GSEA pvalue, GSEA Term, GSEA Count column
organize by Count or pvalue GSEA results (org = 'Count', org = 'pvalue')
num_disp
GSEA pvalue filter
graph title (title)
file_name write out
output visual dotplot of GSEA enrichment
input post GSEA
}
